refactor: CustomView 코드 정리
