+++ app/src/main/java/com/example/slideapp/enums/DrawingType.kt
@@ -1,4 +1,4 @@
-package com.example.slideapp.enum
+package com.example.slideapp.enums
 
 enum class DrawingType {
     IMAGE, SQUARE, DRAW
